+++ apps/enums.py
@@ -1,4 +1,11 @@
 from django.db.models import TextChoices
+from django.utils.translation import gettext_lazy as _
+
+
+class ResponseChoices(TextChoices):
+    CUSTOM = 'CUSTOM', _('Custom Response')
+    PROXY_CURRENT = 'PROXY_CURRENT', _('Proxy Specific URL')
+    PROXY_GLOBAL = 'PROXY_GLOBAL', _('Global Proxy')
 
 
 class HTTPMethods(TextChoices):
@@ -14,4 +21,4 @@ class HTTPMethods(TextChoices):
 class BodyFormat(TextChoices):
     JSON = 'JSON', 'JSON'
     XML = 'XML', 'XML'
-    PLAIN_TEXT = 'PLAIN_TEXT', 'Plain Text'
+    PLAIN_TEXT = 'PLAIN_TEXT', 'Text'