+++ apps/tests/data.py
@@ -1,6 +1,7 @@
 from django.contrib.auth.models import User
 from faker import Faker
 
+from apps.enums import ResponseChoices
 from apps.models import Application, ResourceStub, ResponseStub
 from apps.tests import factories
 
@@ -39,11 +40,12 @@ def create_response_stub(**kwargs) -> ResponseStub:
 
 
 def create_resource_stub(**kwargs) -> ResourceStub:
-    kwargs.setdefault('uri', fake.slug())
+    kwargs.setdefault('slug', fake.slug())
     kwargs.setdefault('description', fake.paragraph(nb_sentences=3))
     kwargs.setdefault('method', fake.http_method())
     kwargs.setdefault('application', create_application())
     kwargs.setdefault('response', create_response_stub())
     kwargs.setdefault('proxy_destination_address', None)
+    kwargs.setdefault('response_type', ResponseChoices.CUSTOM)
 
     return factories.ResourceStubFactory.create(**kwargs)