+++ stubborn/urls.py
@@ -22,7 +22,7 @@ admin.site.site_title = "Stubborn"
 admin.site.index_title = "Administration"
 
 urlpatterns = [
-    path('', RedirectView.as_view(url='/admin')),
-    path('admin', admin.site.urls),
+    path('', RedirectView.as_view(url='/admin/')),
+    path('admin/', admin.site.urls),
     path('', include(('apps.urls', 'apps'), namespace='apps'))
 ]